+++ src/clis/boss/detail.ts
@@ -2,7 +2,7 @@
  * BOSS直聘 job detail — fetch full job posting details via browser cookie API.
  */
 import { cli, Strategy } from '../../registry.js';
-import { requirePage, navigateTo, bossFetch, verbose } from './common.js';
+import { requirePage, navigateTo, bossFetch, verbose } from './utils.js';
 
 cli({
   site: 'boss',
