+++ src/clis/boss/recommend.ts
@@ -2,7 +2,7 @@
  * BOSS直聘 recommend — view recommended candidates (新招呼/greet sort list).
  */
 import { cli, Strategy } from '../../registry.js';
-import { requirePage, navigateToChat, bossFetch, fetchRecommendList, verbose } from './common.js';
+import { requirePage, navigateToChat, bossFetch, fetchRecommendList, verbose } from './utils.js';
 
 cli({
   site: 'boss',
